# src/main/java/com/openbanking/beneficiaries/model/OBReadBeneficiary5.java
package com.openbanking.beneficiaries.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadBeneficiary5
 */
@Getter
@Setter
public class OBReadBeneficiary5 implements Serializable {
	private static final long serialVersionUID = 1L;

}